+++ packages/vector_graphics/lib/vector_graphics_compat.dart
@@ -11,5 +11,6 @@ export 'src/vector_graphics.dart'
         RenderingStrategy,
         VectorGraphic,
         VectorGraphicUtilities,
+        VectorGraphicsImageWidget,
         createCompatVectorGraphic,
         vg;